+++ spec/parser_regex_spec.rb
@@ -2,7 +2,7 @@
 require 'spec_helper'
 
 
-describe Neg::Parser::CharacterParser do
+describe Neg::Parser::RegexParser do
 
   context '_ (any)' do
 
@@ -72,6 +72,39 @@ describe Neg::Parser::CharacterParser do
       parser.to_s.strip.should == %q{
 :
   text == (`tel:` + _("0-9-") * 1)
+      }.strip
+    end
+  end
+
+  context "_(/[0-9-]+/) (direct regex)" do
+
+    let(:parser) {
+      Class.new(Neg::Parser) do
+        text == _(/^tel:[\d-]+/)
+      end
+    }
+
+    it 'parses "tel:0-99"' do
+
+      parser.parse('tel:0-99').should ==
+        [ :text, 0, true, 'tel:0-99', [] ]
+    end
+
+    it 'fails gracefully' do
+
+      parser.parse('tel:a-bc').should ==
+        [ :text,
+          0,
+          false,
+          '"tel:a-bc" doesn\'t match /^tel:[\\d-]+/',
+          [] ]
+    end
+
+    it 'is rendered correctly via #to_s' do
+
+      parser.to_s.strip.should == %q{
+:
+  text == _(/^tel:[\d-]+/)
       }.strip
     end
   end